+++ src/cc4x.hpp
@@ -11,11 +11,11 @@ class cc4x {
    static bool drccd;
    static bool ccsd;
    static bool ref;
-   static int No;
-   static int Nv;
-   static int Nx;
-   static int NF;
-   static int iterations;
+   static size_t No;
+   static size_t Nv;
+   static size_t Nx;
+   static size_t NF;
+   static size_t iterations;
    static CTF::World *dw;
    static kMesh *kmesh;
 };
